Refactor scheduler to initialize lazily in another thread on app start
